Changes to make it compile and remove warnings in VS 2003. (Hopefully it doesn't break it for VS6, I'll find out when i get home).
git-svn-id: http://svn.nate.deepcreek.org.au/svn/DigitalWatch/trunk@91 a123065e-f578-804e-ac06-b1832d43b128

+++ src/AppData.h
@@ -95,8 +95,8 @@ public:
 			struct SETTINGS_ASPECT_RATIO
 			{
 				long bOverride;
-				float width;
-				float height;
+				long width;
+				long height;
 			} aspectRatio;
 
 			int zoom;
@@ -173,8 +173,8 @@ public:
 			struct VALUES_ASPECT_RATIO
 			{
 				long bOverride;
-				float width;
-				float height;
+				long width;
+				long height;
 			} aspectRatio;
 
 			long zoom;
