+++ src/TSFileSource/TSFileStreamList.cpp
@@ -222,7 +222,7 @@ LPWSTR TSFileStreamList::GetListItem(LPWSTR name, long nIndex)
 {
 	CAutoLock listLock(&m_listLock);
 
-	if (nIndex >= m_list.size())
+	if (nIndex >= (long)m_list.size())
 		return NULL;
 
 	long startsWithLength = strStartsWith(name, m_dataListName);
